broke up the screens
