+++ App.js
@@ -5,75 +5,11 @@ import { StyleSheet, Text,  TextInput, View, useColorScheme } from 'react-native
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-
-function EditorScreen() {
-
-  const colorScheme = useColorScheme();
-
-  const themeTextStyle = colorScheme === 'light' ? styles.lightThemeText : styles.darkThemeText;
-  const themeContainerStyle = colorScheme === 'light' ? styles.lightContainer : styles.darkContainer;
-
-  const [text, setText] = useState('');
-  return (
-    <View style={[styles.container, themeContainerStyle]}>
-      <TextInput
-        style={[styles.editor, themeTextStyle]}
-        placeholder="type here"
-        placeholderTextColor="gray"
-
-        defaultValue={text}
-
-        autoComplete="off"
-        autoCorrect={false}
-        autoCapitalize="none"
-        spellCheck={false}
-        textContentType="none"
-
-        multiline={true}
-        numberOfLines={4}
-        adjustsFontSizeToFit={true}
-
-        // autoFocus={true}
-        // clearTextOnFocus={true}
-
-        onChangeText={newText => setText(newText)}
-      />
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
-function HelpScreen() {
-  return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>About Screen</Text>
-    </View>
-  );
-}
-
-function AboutScreen() {
-  return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>About Screen</Text>
-    </View>
-  );
-}
-
-function SettingsScreen() {
-  return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>Settings Screen</Text>
-    </View>
-  );
-}
-
-function HistoryScreen() {
-  return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-      <Text>History Screen</Text>
-    </View>
-  );
-}
+import EditorScreen from './Screens/Editor'
+import HistoryScreen from './Screens/History'
+import SettingsScreen from './Screens/Settings'
+import HelpScreen from './Screens/Help'
+import AboutScreen from './Screens/About'
 
 const Stack = createNativeStackNavigator();
 
@@ -92,35 +28,6 @@ function App() {
   );
 }
 
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  lightContainer: {
-    backgroundColor: '#fff',
-  },
-  darkContainer: {
-    backgroundColor: '#000',
-  },
-  lightThemeText: {
-    color: '#242c40',
-  },
-  darkThemeText: {
-    color: '#fff',
-  },
-  editor: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 10,
-    height: '100%',
-    width: '100%',
-    fontSize: 100,
-    fontWeight: 'bold',
-  }
-});
 
 
 export default App;
