+++ Screens/History.js
@@ -1,0 +1,12 @@
+
+import {  Text,  View } from 'react-native';
+
+function History() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>About Screen</Text>
+    </View>
+  );
+}
+
+export default History
